# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Bike Rental
  description: Bike Rental
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/bike-rental/v1'
...
paths:
...
  /bikes:
    get:
      tags:
        - bike
      summary: Get all bikes
      description: Returns a list of all currently existing bikes
      operationId: getBikes
      parameters:
        - in: query
          name: brand
          schema:
            type: string
          description: Filter result by `brand`
        - in: query
          name: type
          schema:
            $ref: '#/components/schemas/BikeType'
          description: Filter result by `type`
        - in: query
          name: electrified
          schema:
            type: boolean
          description: Filter result by `electrified`
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bike'
  /bikes/{bikeId}:
    get:
      tags:
        - bike
      summary: Get information for one bike
      description: Returns information for bike with `bikeId`
      operationId: getBike
      parameters:
        - in: path
          name: bikeId
          description: Id of bike for which information are requested
          required: true
          schema:
            $ref: '#/components/schemas/BikeId'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bike'
        404:
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - bike
      summary: Remove bike from inventory
      description: Remove a bike with `bikeId` from inventory
      operationId: deleteBike
      parameters:
        - in: path
          name: bikeId
          description: Id of bike for which information are requested
          required: true
          schema:
            $ref: '#/components/schemas/BikeId'
      responses:
        204:
          description: Successful operation
        404:
          $ref: '#/components/responses/NotFound'
  /bikes/{bikeId}/rentals:
    put:
      tags:
        - rental
        - bike
      summary: Rent a bike
      description: Rent the bike `bikeId`
      operationId: rentBike
      parameters:
        - in: path
          name: bikeId
          description: Id of bike which is rented
          required: true
          schema:
            $ref: '#/components/schemas/BikeId'
      requestBody:
        description: Rent bike with `bikeId` to User with `userId`
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
              properties:
                userId:
                  $ref: '#/components/schemas/UserId'
      responses:
        200:
          description: Successfully updated existing rental
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        201:
          description: Successfully created new rental
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rental'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
    BadRequest:
      description: Your request is missing required informations or is ill-formatted
    NotFound:
      description: The representation you are looking for does not exist
  schemas:
...
    Rental:
      type: object
      required:
        - id
        - start
        - startStationId
      properties:
        id:
          $ref: '#/components/schemas/RentalId'
        userId:
          $ref: '#/components/schemas/UserId'
        bikeId:
          $ref: '#/components/schemas/BikeId'
        start:
          type: string
          format: date-time
        startStationId:
          $ref: '#/components/schemas/StationId'
        end:
          type: string
          format: date-time
        endStationId:
          $ref: '#/components/schemas/StationId'
    Bike:
      type: object
      required:
        - id
        - dayOfPurchase
        - stationId
        - brand
        - electrified
        - type
      properties:
        id:
          $ref: '#/components/schemas/BikeId'
        dayOfPurchase:
          type: string
          format: date
        stationId:
          $ref: '#/components/schemas/StationId'
        brand:
          type: string
        electrified:
          type: boolean
          default: false
        type:
          $ref: '#/components/schemas/BikeType'
    BikeId:
      type: string
      format: uuid
    StationId:
      type: string
      format: uuid
    UserId:
      type: string
      format: uuid
    RentalId:
      type: string
      format: uuid
    BikeType:
      type: string
      enum:
        - Mountainbike
        - Gravelbike
        - Trekkingbike
...